pequeños combios

+++ backend-sqrc/src/main/java/com/sqrc/module/backendsqrc/plantillaRespuesta/DTO/PreviewResponseDTO.java
@@ -1,7 +1,7 @@
 package com.sqrc.module.backendsqrc.plantillaRespuesta.DTO;
 
 public record PreviewResponseDTO(
-        Integer templateId,  //revisar luego
+        Long templateId,  //revisar luego
         String titulo,
         String htmlRenderizado
 ) {
